+++ apps/demo/src/app/app.component.html
@@ -1,10 +1,10 @@
 <mat-toolbar color="primary">
   <span>Smart NgRX Demo</span>
   <nav mat-tab-nav-bar [backgroundColor]="'primary'" [tabPanel]="tabPanel">
-    <a mat-tab-link [routerLink]="'tree'" [active]="'tree' === activeLink">
+    <a mat-tab-link [routerLink]="'tree'" [active]="'tree' === activeLink" matTooltip="Tree with standard Mark and Delete">
       Tree
     </a>
-    <a mat-tab-link [routerLink]="'memlab'" [active]="'memlab' === activeLink">
+    <a mat-tab-link [routerLink]="'memlab'" [active]="'memlab' === activeLink" matTooltip="Used to test for memory leaks">
       MemLab
     </a>
   </nav>
