+++ apps/demo/src/app/routes/tree-with-standard-mark-and-delete/components/sidebar.component.html
