+++ apps/demo/src/app/routes/tree-with-standard-mark-and-delete/shared.selectors.ts
@@ -1,6 +1,6 @@
 import { createFeatureSelector } from '@ngrx/store';
 
-import { SharedState, SharedState2 } from './shared-state.interface';
+import { SharedState, SharedState2 } from './store/shared-state.interface';
 
 export const selectSharedState = createFeatureSelector<SharedState>('shared');
 export const selectSharedState2 =
