+++ apps/demo/src/app/routes/tree-with-standard-mark-and-delete/store/current-location/current-location.actions.ts
