+++ apps/demo/src/app/routes/tree-with-standard-mark-and-delete/store/department/children-transform.function.ts
