+++ apps/demo/src/app/routes/tree-with-standard-mark-and-delete/store/department/department.selector.ts
