+++ apps/demo/src/app/routes/tree-with-standard-mark-and-delete/store/docs/doc.interface.ts
