+++ apps/demo/src/app/routes/tree-with-standard-mark-and-delete/store/folders/folders.service.ts
