+++ apps/demo/src/app/routes/tree-with-standard-mark-and-delete/store/functions/add-is-dirty.function.ts
