+++ apps/demo/src/app/routes/tree-with-standard-mark-and-delete/store/lists/lists.service.ts
