+++ apps/demo/src/app/routes/tree-with-standard-mark-and-delete/store/locations/location.interface.ts
