+++ apps/demo/src/app/routes/tree-with-standard-mark-and-delete/store/locations/location.selectors.ts
@@ -4,8 +4,8 @@ import { castTo } from '@smart/smart-ngrx/common/cast-to.function';
 import { createSmartSelector } from '@smart/smart-ngrx/selector/create-smart-selector.function';
 import { MarkAndDeleteSelector } from '@smart/smart-ngrx/types/mark-and-delete-selector.type';
 
-import { selectCurrentLocationId } from '../../current-location/current-location.selector';
 import { selectSharedState } from '../../shared.selectors';
+import { selectCurrentLocationId } from '../current-location/current-location.selector';
 import { selectDepartmentsChildren } from '../department/department.selector';
 import { Location } from './location.interface';
