+++ apps/demo/src/app/routes/tree-with-standard-mark-and-delete/store/sprint-folders/sprint-folder.interface.ts
